+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- File-type icons in the explorer (per language/extension; open/closed folders).
+
 ## [0.2.2] - 2026-06-27
 
 ### Added